requirements.txt
Flask
PyJWT
Werkzeug
SQLAlchemy
flask_sqlalchemy
python-dotenv
psycopg2